Notifications: extract SupersededPhoneDismissBuffer to its own file
File-organization policy: the per-tab/surface superseded-banner stash was a
second major type appended to TerminalNotificationStore.swift. Move it to
SupersededPhoneDismissBuffer.swift (no behavior change) and wire it into the
cmux target.
